Some more changes
Changed some structural pieces.
Added a WORKING batch file :))
Added an 'under construction' page.
Not satisfied but has a working home page.
Tiles are still not majorly reactive but this is only a branch for me to work out the design.
Added more server get ports.

remember... it's not a bug, it's a feature.
Patched some annoyances.